# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CareLink API
  version: 0.0.1
  description: >-
    Caregiver slice 1 is implemented for GET /caregivers/me, GET /caregivers/me/schedule,
    and GET /visits/{visitId}/work-pack; these definitions supersede the matching draft operations.
    Includes the existing Auth contract and the specified FM01-FM05 family contracts. Family
    operations remain x-status: draft until implemented; inclusion in Final does not imply a working
    backend. Shared paths retain their non-family response alternatives. The family owner supplies
    only the family use-case behaviour; other use cases, binding management and platform
    implementation are outside this change.
  license:
    name: UNLICENSED
servers:
  - url: /api
    description: Same-origin, relative to whichever host serves the front end.
...
paths:
...
  /visits/{visitId}/check-in:
    parameters:
      - $ref: '#/components/parameters/VisitId'
    post:
      tags: [Visits]
      summary: Record arrival and start assigned service (CG03 basics)
      operationId: caregiverVisitCheckIn
      x-status: implemented
      security: [{sessionCookie: []}]
      description: >-
        CAREGIVER only, CSRF required. Within configured early arrival (default 30 minutes) through scheduledEnd
        or Singapore day end. One transaction initializes only assigned plan-version tasks and records two APPLIED
        transitions SCHEDULED to ARRIVED to IN_PROGRESS. checkedInAt is server time. Manual location is explicitly
        unverified; GPS is device-supplied and no geofence is applied. No check-out or SYS03 scheduler is implemented.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CaregiverCheckInCommand'}
      responses:
        '200':
          description: Saved command result, or original result on identical replay; reload work-pack for current state.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CaregiverExecutionResult'}
        '400': {description: Invalid location or command fields.}
        '401': {description: Authentication required.}
        '403': {description: 'CSRF, role or current assignment rejected.'}
        '404': {description: Visit not found.}
        '409': {description: 'Window, plan, state, version or command conflict. Authorized rejected state attempts are logged.'}
        '503': {description: Storage unavailable; check results before retry.}
  /visits/{visitId}/tasks/{taskId}/complete:
    parameters:
      - $ref: '#/components/parameters/VisitId'
      - {name: taskId, in: path, required: true, schema: {type: integer, format: int64}}
    post:
      tags: [Visits]
      summary: Record an assigned task result (CG03 basics)
      operationId: caregiverVisitTaskResult
      x-status: implemented
      security: [{sessionCookie: []}]
      description: >-
        CAREGIVER only, CSRF required, current assignment and IN_PROGRESS required. PENDING can become DONE,
        SKIPPED or REFUSED once. Only DONE sets completedAt and counts as completed; other results require a
        factual reason and retain null completedAt. Every write advances the parent Visit version. No task result
        completes the Visit. Identical successful commands replay without new timestamps; parent/task ownership
        is checked server-side, not selected through a client projection parameter.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CaregiverTaskResultCommand'}
      responses:
        '200':
          description: Original successful command result; reload tasks/work-pack for current state.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CaregiverExecutionResult'}
        '400': {description: Invalid result or missing reason.}
        '401': {description: Authentication required.}
        '403': {description: 'Role, CSRF or assignment rejected.'}
        '404': {description: Visit or task not found in this visit.}
        '409': {description: 'State, version, task terminal result or command conflict.'}
        '503': {description: Storage unavailable; check results before retry.}
...
  /visits:
    get:
      tags:
        - Visits
      summary: List visits
      x-status: draft
      x-family-status: implemented
      description: >-
        The implemented FAMILY list reads visits for elders currently accessible to the caller.
        Identity comes from the authenticated session; the account must be enabled, hold FAMILY
        and have a family profile. Each request requires an ACTIVE, unexpired elder binding with
        FULL or READ_ONLY scope. Expiry is evaluated in Asia/Singapore and access ends exactly at
        expiresAt. No familyMemberId or role parameter is accepted. This read-only GET requires
        no CSRF header.

        Omit elderId to include all currently readable elders. An explicitly inaccessible elderId
        returns 403. A caregiverId filter requires a current or historical visit whose caregiverId
        matches and whose elder is currently readable. This relationship check uses all readable
        elders, independently of the requested date range or selected elder; a replacement-history
        row alone grants no access. An unrelated caregiverId returns 403.

        Provide dateFrom and dateTo together, or omit both to use the current Monday-to-Sunday
        week in Asia/Singapore. Dates are inclusive: scheduledStart must be at or after dateFrom
        00:00 and before the day after dateTo 00:00. Both dates must be within 1000-01-01 through
        9999-12-30; a missing date, reversed range, malformed date, invalid status, nonpositive
        resource ID or invalid pagination returns 400. Authorization and all date, status, elder
        and caregiver filters apply before counting and pagination. Sort by scheduledStart then
        id ascending. Page defaults to 0 and size to 20; page must be nonnegative and size 1-200.
        No readable elders, no matching visits and pages beyond the last page return 200 with
        items=[]. totalElements counts all authorized matches before pagination.

        Example: GET /api/visits?elderId=101&dateFrom=2026-09-21&dateTo=2026-09-27&page=0&size=2.
        Null caregiverId means unassigned; null serviceType means the source visit has no service
        type. Timestamps include an explicit offset. All items share the server snapshot asOf
        for this response.

        Other-role contracts remain planned: managers can filter freely, caregivers see their
        own visits, and elders see their own scope. UC-CG06 uses a past dateTo for historical
        records. Their original response alternative is retained; their collection-query
        implementation is pending and non-FAMILY callers currently receive 403.

        Validated requests reaching the FAMILY application query persist a READ record with
        the actor app_user ID (when resolvable), resource and outcome: OK, including empty
        results; DENIED for access refusal; or FAILED for lookup/query errors. Records
        commit independently with server occurred_at in Asia/Singapore. Authentication 401,
        method-role rejection 403 and MVC validation/conversion 400 rejected before the
        FAMILY application query are outside this audit. Existing request logging is
        unchanged. Audit storage failure returns 503 without protected data. Records
        exclude response bodies, raw query strings, exception details and secrets. Reading
        does not acknowledge notifications or incidents. The resource is VISIT with a null
        collection ID. Audit detail contains only validated elder/caregiver IDs, resolved
        inclusive dates, status, page and size.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: elderId
          in: query
          description: Optional readable elder ID. Omit to query all currently readable elders.
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 101
        - name: caregiverId
          in: query
          description: Optional caregiver ID with a visit relationship to a currently readable elder.
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 201
        - name: dateFrom
          in: query
          description: Inclusive Singapore start date, 1000-01-01 through 9999-12-30. Requires dateTo.
          schema:
            type: string
            format: date
          example: '2026-09-21'
        - name: dateTo
          in: query
          description: Inclusive Singapore end date, 1000-01-01 through 9999-12-30. Requires dateFrom.
          schema:
            type: string
            format: date
          example: '2026-09-27'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/VisitStatus'
          example: SCHEDULED
      responses:
        '200':
          description: >-
            The implemented family list uses FamilyVisitPage. The original response alternative
            is retained for other roles, whose collection-query implementation is pending.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/FamilyVisitPage'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Visit'
                      page:
                        type: integer
                      size:
                        type: integer
                      totalElements:
                        type: integer
              examples:
                familyWeek:
                  summary: First page with assigned and unassigned visits
                  value:
                    items:
                      - id: 301
                        elderId: 101
                        caregiverId: 201
                        serviceType: BATHING
                        scheduledStart: '2026-09-24T09:00:00+08:00'
                        scheduledEnd: '2026-09-24T10:00:00+08:00'
                        checkedInAt: null
                        checkedOutAt: null
                        status: SCHEDULED
                        asOf: '2026-09-23T10:00:00+08:00'
                      - id: 302
                        elderId: 101
                        caregiverId: null
                        serviceType: null
                        scheduledStart: '2026-09-25T09:00:00+08:00'
                        scheduledEnd: null
                        checkedInAt: null
                        checkedOutAt: null
                        status: SCHEDULED
                        asOf: '2026-09-23T10:00:00+08:00'
                    page: 0
                    size: 2
                    totalElements: 3
                empty:
                  summary: No readable elders or no matching visits
                  value:
                    items: []
                    page: 0
                    size: 20
                    totalElements: 0
                beyondLastPage:
                  summary: Requested page is past three authorized matches
                  value:
                    items: []
                    page: 2
                    size: 2
                    totalElements: 3
        '400':
          description: Invalid resource ID, date range, status or pagination.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: >-
            No valid authenticated session. Shared security determines the response body;
            clients must handle the HTTP status without requiring a JSON body.
        '403':
          description: >-
            The caller lacks FAMILY, the current family account/profile is unavailable, an
            explicit elderId is not readable, or caregiverId has no permitted visit relationship.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '503':
          $ref: '#/components/responses/FamilyAuditUnavailable'
      x-contract-status: specified
      x-family-use-cases:
        - UC-FM02
        - UC-FM03
      x-family-owner: Member 4 (FM01-FM05 behaviour); source data remains with its assigned owner.
      security:
        - sessionCookie: []
      operationId: listVisits
      x-family-response-schema:
        $ref: '#/components/schemas/FamilyVisitPage'
...
components:
  schemas:
...
    CaregiverCommandIdentity:
      type: object
      required: [expectedVersion, clientRequestId]
      properties:
        expectedVersion: {type: integer, minimum: 0}
        clientRequestId: {type: string, format: uuid, description: Immutable command key; retain original payload and key when the result is unknown.}
...
    CaregiverCheckInCommand:
      allOf:
        - $ref: '#/components/schemas/CaregiverCommandIdentity'
        - type: object
          required: [locationSource]
          description: GPS requires coordinates and nonnegative finite accuracy, without note; MANUAL_LOCATION_NOTE requires a nonblank note without GPS fields.
          properties:
            locationSource: {type: string, enum: [GPS, MANUAL_LOCATION_NOTE]}
            latitude: {type: number, minimum: -90, maximum: 90}
            longitude: {type: number, minimum: -180, maximum: 180}
            accuracy: {type: number, minimum: 0}
            locationNote: {type: string, maxLength: 500}
            clientCapturedAt: {type: string, format: date-time, description: Optional ISO instant; cannot override server check-in time.}
    CaregiverTaskResultCommand:
      allOf:
        - $ref: '#/components/schemas/CaregiverCommandIdentity'
        - type: object
          required: [status]
          properties:
            status: {type: string, enum: [DONE, SKIPPED, REFUSED]}
            outcome: {type: string, maxLength: 255, nullable: true}
            caregiverNote: {type: string, maxLength: 500, nullable: true, description: Nonblank factual reason required for SKIPPED or REFUSED.}
...
    CaregiverExecutionResult:
      type: object
      required: [visitId, visitVersion, savedState, taskId, replayed]
      properties:
        visitId: {type: integer, format: int64}
        visitVersion: {type: integer, description: 'Version of the original command, not necessarily current version on replay.'}
        savedState: {type: string, nullable: true, description: Original saved state; null for a task replay.}
        taskId: {type: integer, format: int64, nullable: true}
        replayed: {type: boolean}
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string

    # Family projections and referenced shared schemas.
...
    VisitStatus:
      type: string
      enum:
      - SCHEDULED
      - ARRIVED
      - IN_PROGRESS
      - COMPLETED
      - VERIFIED
      - AUTO_CLOSED
      - EXCEPTION
      - CANCELLED
    FamilyVisitPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FamilyVisitView'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
          maximum: 200
        totalElements:
          type: integer
          format: int64
          minimum: 0
      required:
        - items
        - page
        - size
        - totalElements
    FamilyVisitView:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        elderId:
          type: integer
          format: int64
        caregiverId:
          type: integer
          format: int64
          nullable: true
        serviceType:
          type: string
          nullable: true
          description: Source visit service type; null when not recorded.
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
          nullable: true
        checkedInAt:
          type: string
          format: date-time
          nullable: true
        checkedOutAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/VisitStatus'
        asOf:
          type: string
          format: date-time
          description: >-
            Server snapshot time for this read, shared by every item in the response; not the
            last modification time. A failed refresh must not advance it.
      description: >-
        Family service view. No absence record, internal state deadline, or care-plan implementation
        IDs. Null caregiverId means unassigned. Timestamps are RFC 3339 values with an explicit
        offset; visit times are interpreted in Asia/Singapore. Nullable source fields remain null.
      required:
        - id
        - elderId
        - caregiverId
        - serviceType
        - scheduledStart
        - scheduledEnd
        - checkedInAt
        - checkedOutAt
        - status
        - asOf
    Visit:
      type: object
      required:
      - id
      - elderId
      - status
      - scheduledStart
      properties:
        id:
          type: integer
          format: int64
        elderId:
          type: integer
          format: int64
        caregiverId:
          type: integer
          format: int64
          nullable: true
        carePlanNodeId:
          type: integer
          format: int64
          nullable: true
        carePlanId:
          type: integer
          format: int64
          nullable: true
          description: >-
            The care plan VERSION in force when this visit was created — a snapshot by reference, since superseded
            versions are never edited.
        absenceId:
          type: integer
          format: int64
          nullable: true
        serviceType:
          type: string
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
          nullable: true
        checkedInAt:
          type: string
          format: date-time
          nullable: true
        checkedOutAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/VisitStatus'
        stateDeadline:
          type: string
          format: date-time
          nullable: true
          description: >-
            Optional deadline for other visit states. SYS03 uses scheduledStart plus the configured
            late-arrival threshold, not this field; its alert transitions SCHEDULED to EXCEPTION and blocks check-in,
            including refreshed requests. Resolving the Incident does not reopen the Visit.
        assignmentHistory:
          type: array
          description: Previous and current caregiver assignments, including replacement status, reason and assignment times.
          items:
            $ref: '#/components/schemas/VisitAssignmentHistoryItem'
...
    VisitAssignmentHistoryItem:
      type: object
      properties:
        caregiverId:
          type: integer
          format: int64
        caregiverName:
          type: string
        status:
          type: string
        reason:
          type: string
          nullable: true
        assignedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
...
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
    VisitId:
      name: visitId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
...
    FamilyAuditUnavailable:
      description: >-
        The family access record could not be committed. After any query outcome, failure to
        record access returns 503 instead of the result or original error, without protected data.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          example:
            title: Access recording unavailable
            status: 503
            detail: Access could not be recorded. Please retry later.
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: >
        Server-side session established by `POST /auth/login`. State-changing requests must also echo
        the `XSRF-TOKEN` cookie value in an `X-XSRF-TOKEN` header.